Refactor testing fakes out into browser_test_utils.ts
